requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.6
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
